Added BottomNavigation on ListDetailsPage
