+++ lib/presentation/core/localization/es.json
@@ -390,5 +390,8 @@
   "settings_daily_pace": "Ritmo controlado",
   "settings_daily_pace_sub": "Si está activo, podrás realizar 1 prueba al día por cada modo de estudio con un número precalculado de palabras. Si no, no hay restricciones",
   "filter_recently_added": "Reciente",
-  "test_info_controlled_pace": "Solo podrás completar un test diario por modo de estudio al día."
+  "test_info_controlled_pace": "Solo podrás completar un test diario por modo de estudio al día.",
+  "list_details_words": "Palabras",
+  "list_details_grammar": "Gramática",
+  "list_details_searchBar_hint_grammar": "Buscar por nombre, definición..."
 }
