+++ lib/presentation/list_details_page/list_details_page.dart
@@ -1,31 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:kanpractice/application/list_details/list_details_bloc.dart';
-import 'package:kanpractice/application/services/preferences_service.dart';
 import 'package:kanpractice/injection.dart';
 import 'package:kanpractice/presentation/core/routing/pages.dart';
-import 'package:kanpractice/presentation/core/util/tutorial_coach.dart';
-import 'package:kanpractice/presentation/core/types/test_modes.dart';
-import 'package:kanpractice/presentation/core/types/coach_tutorial_parts.dart';
-import 'package:kanpractice/presentation/core/types/study_modes.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:kanpractice/domain/list/list.dart';
-import 'package:kanpractice/domain/word/word.dart';
 import 'package:kanpractice/presentation/add_word_page/arguments.dart';
+import 'package:kanpractice/presentation/core/types/coach_tutorial_parts.dart';
+import 'package:kanpractice/presentation/core/types/list_details_types.dart';
 import 'package:kanpractice/presentation/core/ui/blitz/kp_blitz_bottom_sheet.dart';
 import 'package:kanpractice/presentation/core/ui/folder_list_bottom_sheet.dart';
 import 'package:kanpractice/presentation/core/ui/kp_alert_dialog.dart';
-import 'package:kanpractice/presentation/core/ui/kp_button.dart';
-import 'package:kanpractice/presentation/core/ui/kp_empty_list.dart';
-import 'package:kanpractice/presentation/core/ui/kp_progress_indicator.dart';
 import 'package:kanpractice/presentation/core/ui/kp_scaffold.dart';
 import 'package:kanpractice/presentation/core/ui/kp_search_bar.dart';
 import 'package:kanpractice/presentation/core/ui/kp_text_form.dart';
-import 'package:kanpractice/presentation/core/util/consts.dart';
-import 'package:kanpractice/presentation/core/util/utils.dart';
-import 'package:kanpractice/presentation/list_details_page/widgets/kanji_item.dart';
-import 'package:kanpractice/presentation/list_details_page/widgets/practice_on_list_bottom_sheet.dart';
-import 'package:kanpractice/presentation/study_modes/utils/mode_arguments.dart';
+import 'package:kanpractice/presentation/core/util/tutorial_coach.dart';
+import 'package:kanpractice/presentation/list_details_page/grammar_list.dart';
+import 'package:kanpractice/presentation/list_details_page/widgets/list_details_bottom_navigation.dart';
+import 'package:kanpractice/presentation/list_details_page/word_list.dart';
 
 class ListDetailsPage extends StatefulWidget {
   final WordList list;
@@ -37,17 +29,15 @@ class ListDetailsPage extends StatefulWidget {
 
 class _ListDetailsPageState extends State<ListDetailsPage>
     with SingleTickerProviderStateMixin {
-  final ScrollController _scrollController = ScrollController();
-
   /// Tutorial Global Keys
   final GlobalKey vocabulary = GlobalKey();
   final GlobalKey actions = GlobalKey();
   final GlobalKey changeName = GlobalKey();
+  final GlobalKey navigation = GlobalKey();
 
-  FocusNode? _searchBarFn;
-  TabController? _tabController;
-  StudyModes _selectedMode = StudyModes.writing;
-  bool _aggrStats = false;
+  final _searchBarFn = FocusNode();
+  final _pageController = PageController();
+  final _searchTextController = TextEditingController();
 
   /// Saves the last state of the query
   String _query = "";
@@ -57,101 +47,24 @@ class _ListDetailsPageState extends State<ListDetailsPage>
 
   bool _searchHasFocus = false;
   bool _onTutorial = false;
+  ListDetailsType _currentPage = ListDetailsType.words;
 
   @override
   void initState() {
-    _searchBarFn = FocusNode();
-    _tabController =
-        TabController(length: StudyModes.values.length, vsync: this);
-    _tabController?.addListener(_tabControllerManagement);
-    _searchBarFn?.addListener(_focusListener);
-    _scrollController.addListener(_scrollListener);
+    _searchBarFn.addListener(_focusListener);
     _listName = widget.list.name;
-    _aggrStats = getIt<PreferencesService>()
-        .readData(SharedKeys.kanListListVisualization);
-    getIt<ListDetailBloc>().add(
-      ListDetailEventLoading(_listName, reset: true),
-    );
     super.initState();
   }
 
   @override
   void dispose() {
-    _searchBarFn?.removeListener(_focusListener);
-    _searchBarFn?.dispose();
-    _tabController?.dispose();
-    _scrollController.removeListener(_scrollListener);
-    _scrollController.dispose();
+    _searchBarFn.removeListener(_focusListener);
+    _searchBarFn.dispose();
+    _pageController.dispose();
     super.dispose();
   }
 
-  _scrollListener() {
-    /// When reaching last pixel of the list
-    if (_scrollController.offset ==
-        _scrollController.position.maxScrollExtent) {
-      /// If the query is empty, use the pagination for search bar
-      if (_query.isNotEmpty) {
-        _addSearchingEvent(_query);
-      }
-
-      /// Else use the normal pagination
-      else {
-        _addLoadingEvent();
-      }
-    }
-  }
-
-  _tabControllerManagement() {
-    _searchBarFn?.unfocus();
-    setState(() {
-      _selectedMode = StudyModes.values[_tabController?.index ?? 0];
-    });
-  }
-
-  _focusListener() =>
-      setState(() => _searchHasFocus = (_searchBarFn?.hasFocus ?? false));
-
-  _onModeChange(StudyModes newMode) {
-    _searchBarFn?.unfocus();
-    setState(() {
-      _selectedMode = newMode;
-    });
-    _tabController?.animateTo(newMode.index,
-        duration: const Duration(milliseconds: KPAnimations.ms300),
-        curve: Curves.easeInOut);
-  }
-
-  _updateSelectedModePageView(double pv) {
-    switch (_selectedMode) {
-      case StudyModes.writing:
-        if (pv < 0) _onModeChange(StudyModes.reading);
-        break;
-      case StudyModes.reading:
-        if (pv < 0) {
-          _onModeChange(StudyModes.recognition);
-        } else if (pv > 0) {
-          _onModeChange(StudyModes.writing);
-        }
-        break;
-      case StudyModes.recognition:
-        if (pv < 0) {
-          _onModeChange(StudyModes.listening);
-        } else if (pv > 0) {
-          _onModeChange(StudyModes.reading);
-        }
-        break;
-      case StudyModes.listening:
-        if (pv < 0) {
-          _onModeChange(StudyModes.speaking);
-        } else if (pv > 0) {
-          _onModeChange(StudyModes.recognition);
-        }
-        break;
-      case StudyModes.speaking:
-        if (pv > 0) _onModeChange(StudyModes.listening);
-        break;
-    }
-  }
+  _focusListener() => setState(() => _searchHasFocus = _searchBarFn.hasFocus);
 
   _addLoadingEvent({bool reset = false}) {
     return getIt<ListDetailBloc>()
@@ -195,20 +108,6 @@ class _ListDetailsPageState extends State<ListDetailsPage>
         });
   }
 
-  Future<void> _goToPractice(ListDetailStateLoadedPractice state) async {
-    await Navigator.of(context)
-        .pushNamed(state.mode.page,
-            arguments: ModeArguments(
-                studyList: state.list,
-                isTest: false,
-                mode: state.mode,
-                testMode: Tests.blitz,
-                studyModeHeaderDisplayName: _listName))
-        .then(
-          (value) => _addLoadingEvent(reset: true),
-        );
-  }
-
   @override
   Widget build(BuildContext context) {
     return KPScaffold(
@@ -216,7 +115,7 @@ class _ListDetailsPageState extends State<ListDetailsPage>
         if (_onTutorial) return false;
         if (_searchHasFocus) {
           _addLoadingEvent(reset: true);
-          _searchBarFn?.unfocus();
+          _searchBarFn.unfocus();
           return false;
         } else {
           return true;
@@ -271,10 +170,24 @@ class _ListDetailsPageState extends State<ListDetailsPage>
           ],
         ),
       ],
+      bottomNavigationWidget: ListDetailsBottomNavigation(
+        key: navigation,
+        currentPage: _currentPage,
+        onPageChanged: (type) {
+          setState(() => _currentPage = type);
+          _searchBarFn.unfocus();
+          _searchTextController.text = "";
+          _pageController.jumpToPage(type.page);
+        },
+      ),
       child: Column(
         children: [
+          // TODO: Grammar queries
           KPSearchBar(
-            hint: "list_details_searchBar_hint".tr(),
+            controller: _searchTextController,
+            hint: _currentPage == ListDetailsType.words
+                ? "list_details_searchBar_hint".tr()
+                : "list_details_searchBar_hint_grammar".tr(),
             focus: _searchBarFn,
             onQuery: (String query) {
               /// Everytime the user queries, reset the query itself and
@@ -289,141 +202,28 @@ class _ListDetailsPageState extends State<ListDetailsPage>
             },
           ),
           Expanded(
-            child: BlocConsumer<ListDetailBloc, ListDetailState>(
-              key: vocabulary,
-              listener: (context, state) async {
-                if (state is ListDetailStateLoadedPractice) {
-                  await _goToPractice(state);
-                } else if (state is ListDetailStateFailure) {
-                  if (state.error.isNotEmpty) {
-                    Utils.getSnackBar(context, state.error);
-                  }
-                } else if (state is ListDetailStateLoaded) {
-                  if (getIt<PreferencesService>().readData(
-                          SharedKeys.haveSeenKanListDetailCoachMark) ==
-                      false) {
+            child: PageView(
+              controller: _pageController,
+              children: [
+                WordListWidget(
+                  key: vocabulary,
+                  list: widget.list,
+                  query: _query,
+                  searchBarFn: _searchBarFn,
+                  onStartTutorial: () async {
                     _onTutorial = true;
                     await TutorialCoach([vocabulary, actions, changeName],
                             CoachTutorialParts.details)
                         .showTutorial(context,
                             onEnd: () => _onTutorial = false);
-                  }
-                }
-              },
-              builder: (context, state) {
-                if (state is ListDetailStateLoaded) {
-                  _listName = state.name;
-                  return _body(state);
-                } else if (state is ListDetailStateLoading ||
-                    state is ListDetailStateSearching ||
-                    state is ListDetailStateIdle ||
-                    state is ListDetailStateLoadedPractice) {
-                  return const KPProgressIndicator();
-                } else if (state is ListDetailStateFailure) {
-                  return KPEmptyList(
-                      showTryButton: true,
-                      onRefresh: () => _addLoadingEvent(reset: true),
-                      message: "list_details_load_failed".tr());
-                } else {
-                  return Container();
-                }
-              },
+                  },
+                ),
+                const GrammarListWidget(),
+              ],
             ),
           ),
         ],
       ),
-    );
-  }
-
-  Column _body(ListDetailStateLoaded state) {
-    return Column(
-      children: [
-        if (!_aggrStats)
-          Padding(
-            padding: const EdgeInsets.only(bottom: KPMargins.margin8),
-            child: TabBar(
-                controller: _tabController,
-                tabs: List.generate(StudyModes.values.length, (index) {
-                  return Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 8.0),
-                    child: Icon(
-                      StudyModes.values[index].icon,
-                      color: StudyModes.values[index].color,
-                    ),
-                  );
-                })),
-          ),
-        _aggrStats
-            ? Expanded(child: _kanjiList(state))
-            : Expanded(
-                child: GestureDetector(
-                /// Dismiss keyboard if possible whenever a vertical or horizontal
-                /// drag down occurs on screen
-                onVerticalDragStart: (details) {
-                  FocusManager.instance.primaryFocus?.unfocus();
-                },
-                onHorizontalDragStart: (details) {
-                  FocusManager.instance.primaryFocus?.unfocus();
-                },
-                onHorizontalDragEnd: (details) {
-                  double? pv = details.primaryVelocity;
-                  if (pv != null) _updateSelectedModePageView(pv);
-                },
-                child: _kanjiList(state),
-              )),
-        KPButton(
-            title1: "list_details_practice_button_label_ext".tr(),
-            title2: "list_details_practice_button_label".tr(),
-            onTap: () async {
-              if (_aggrStats) {
-                return await PracticeListBottomSheet.show(
-                        context, _listName, state.list)
-                    .then(
-                  (value) => _addLoadingEvent(reset: true),
-                );
-              }
-              getIt<ListDetailBloc>()
-                  .add(ListDetailEventLoadUpPractice(_listName, _selectedMode));
-            }),
-      ],
-    );
-  }
-
-  Widget _kanjiList(ListDetailStateLoaded state) {
-    if (state.list.isEmpty) {
-      return KPEmptyList(
-          showTryButton: true,
-          onRefresh: () => _addLoadingEvent(reset: true),
-          message: "list_details_empty".tr());
-    }
-    return GridView.builder(
-      key: const PageStorageKey<String>('kanjiListController'),
-      itemCount: state.list.length,
-      controller: _scrollController,
-      keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-          crossAxisCount: 5, childAspectRatio: 2),
-      itemBuilder: (context, k) {
-        Word? kanji = state.list[k];
-        return KanjiItem(
-          aggregateStats: _aggrStats,
-          index: k,
-          kanji: kanji,
-          list: widget.list,
-          listName: _listName,
-          selectedMode: _selectedMode,
-          onShowModal: () => _searchBarFn?.unfocus(),
-          onTap: () async {
-            await Navigator.of(context)
-                .pushNamed(KanPracticePages.addKanjiPage,
-                    arguments: AddWordArgs(listName: _listName, word: kanji))
-                .then((code) {
-              if (code == 0) _addLoadingEvent(reset: true);
-            });
-          },
-          onRemoval: () => _addLoadingEvent(reset: true),
-        );
-      },
     );
   }
 }
